Move selection networks to networks package.
