+++ nengo_spa/networks/tests/test_selection.py
@@ -2,7 +2,7 @@ import numpy as np
 from numpy.testing import assert_allclose
 
 import nengo
-from nengo_spa.selection import IA, Thresholding, WTA
+from nengo_spa.networks.selection import IA, Thresholding, WTA
 
 
 def test_ia(Simulator, plt, seed):
